refactor(shipping): Phase 4 split EfShipmentStore from ShippingContext
- New EfShipmentStore implements IShipmentStore via injected
  ShippingContext; DI flipped from ShippingContext to EfShipmentStore.
- ShippingContext reduced to DbContext + DbSets + OnModelCreating;
  IShipmentStore interface and 8 method bodies removed.
- Single commit covers all 3 sub-steps per issue #214 to avoid a
  misleading 'two impls coexist' state on bisect.
