+++ shipping-microservice/Shipping.Service/Infrastructure/Data/EntityFramework/ShippingContext.cs
@@ -1,10 +1,9 @@
 using Microsoft.EntityFrameworkCore;
 using Shipping.Service.Domain;
-using Shipping.Service.Domain.Abstractions;
 
 namespace Shipping.Service.Infrastructure.Data.EntityFramework;
 
-internal class ShippingContext : DbContext, IShipmentStore
+internal class ShippingContext : DbContext
 {
     public ShippingContext(DbContextOptions<ShippingContext> options)
         : base(options)
@@ -24,141 +23,5 @@ internal class ShippingContext : DbContext, IShipmentStore
         modelBuilder.ApplyConfiguration(new ShipmentLineConfiguration());
         modelBuilder.ApplyConfiguration(new ShipmentStatusHistoryEntryConfiguration());
         modelBuilder.ApplyConfiguration(new OrderConfirmationConfiguration());
-    }
-
-    public async Task<IReadOnlyList<Shipment>> GetByOrder(Guid orderId)
-    {
-        return await Shipments
-            .Include(s => s.Lines)
-            .Where(s => s.OrderId == orderId)
-            .OrderBy(s => s.WarehouseId)
-            .ToListAsync();
-    }
-
-    public async Task<Shipment?> GetById(Guid shipmentId)
-    {
-        return await Shipments
-            .Include(s => s.Lines)
-            .FirstOrDefaultAsync(s => s.Id == shipmentId);
-    }
-
-    public async Task<IReadOnlyList<Shipment>> ListShipments(ShipmentListFilters filters)
-    {
-        var query = Shipments
-            .Include(s => s.Lines)
-            .AsQueryable();
-
-        if (filters.Status is not null)
-        {
-            query = query.Where(s => s.Status == filters.Status);
-        }
-
-        if (filters.WarehouseId is not null)
-        {
-            query = query.Where(s => s.WarehouseId == filters.WarehouseId);
-        }
-
-        if (filters.From is not null)
-        {
-            query = query.Where(s => s.CreatedAt >= filters.From);
-        }
-
-        if (filters.To is not null)
-        {
-            query = query.Where(s => s.CreatedAt <= filters.To);
-        }
-
-        return await query
-            .OrderByDescending(s => s.CreatedAt)
-            .Skip(filters.Skip)
-            .Take(filters.Take)
-            .ToListAsync();
-    }
-
-    public Task<int> SaveChangesAsync() => base.SaveChangesAsync();
-
-    public async Task<IReadOnlyList<Shipment>> ListActiveShipments()
-    {
-        return await Shipments
-            .Include(s => s.Lines)
-            .Where(s => s.Status == ShipmentStatus.Shipped || s.Status == ShipmentStatus.InTransit)
-            .ToListAsync();
-    }
-
-    public async Task<Shipment?> GetByTrackingNumber(string trackingNumber)
-    {
-        if (string.IsNullOrWhiteSpace(trackingNumber))
-        {
-            return null;
-        }
-
-        return await Shipments
-            .Include(s => s.Lines)
-            .FirstOrDefaultAsync(s => s.TrackingNumber == trackingNumber);
-    }
-
-    public async Task<CreateShipmentsResult> CreateShipmentsForOrder(
-        Guid orderId,
-        string customerId,
-        IReadOnlyList<CreateShipmentLine> lines)
-    {
-        var existing = await Shipments
-            .AnyAsync(s => s.OrderId == orderId);
-
-        if (existing)
-        {
-            return new CreateShipmentsResult(Created: false, Shipments: []);
-        }
-
-        var now = DateTime.UtcNow;
-        var shipments = new List<Shipment>();
-
-        foreach (var group in lines.GroupBy(l => l.WarehouseId))
-        {
-            var shipment = Shipment.Create(
-                id: Guid.NewGuid(),
-                orderId: orderId,
-                customerId: customerId,
-                warehouseId: group.Key,
-                createdAt: now);
-
-            foreach (var line in group)
-            {
-                shipment.AddLine(line.ProductId, line.Quantity);
-            }
-
-            Shipments.Add(shipment);
-            shipments.Add(shipment);
-        }
-
-        await SaveChangesAsync();
-
-        return new CreateShipmentsResult(Created: true, Shipments: shipments);
-    }
-
-    public async Task RecordOrderConfirmation(Guid orderId, string customerId)
-    {
-        var exists = await OrderConfirmations.AnyAsync(o => o.OrderId == orderId);
-        if (exists)
-        {
-            return;
-        }
-
-        OrderConfirmations.Add(new OrderConfirmation
-        {
-            OrderId = orderId,
-            CustomerId = customerId,
-            ReceivedAt = DateTime.UtcNow,
-        });
-
-        await SaveChangesAsync();
-    }
-
-    public async Task<string?> TryGetOrderCustomer(Guid orderId)
-    {
-        var confirmation = await OrderConfirmations
-            .FirstOrDefaultAsync(o => o.OrderId == orderId);
-
-        return confirmation?.CustomerId;
     }
 }
